+++ app/src/main/kotlin/org/fossify/contacts/activities/ViewContactActivity.kt
@@ -2,14 +2,12 @@ package org.fossify.contacts.activities
 
 import android.content.ActivityNotFoundException
 import android.content.ContentUris
-import android.content.Context
 import android.content.Intent
 import android.media.AudioManager
 import android.media.RingtoneManager
 import android.net.Uri
 import android.os.Bundle
 import android.provider.ContactsContract
-import android.telephony.PhoneNumberUtils
 import android.view.View
 import android.view.WindowManager
 import android.widget.RelativeLayout
@@ -103,7 +101,7 @@ class ViewContactActivity : ContactActivity() {
 
             findItem(R.id.edit).setOnMenuItemClickListener {
                 if (contact != null) {
-                    launchEditContact(contact!!)
+                    launchEditContact(contact!!, false)
                 }
                 true
             }
@@ -258,7 +256,7 @@ class ViewContactActivity : ContactActivity() {
             }
         }
 
-        getDuplicateContacts {
+        initializeDuplicateContacts {
             duplicateInitialized = true
             setupContactDetails()
         }
@@ -283,10 +281,14 @@ class ViewContactActivity : ContactActivity() {
         updateTextColors(binding.contactScrollview)
     }
 
-    private fun launchEditContact(contact: Contact) {
+    private fun launchEditContact(contact: Contact, editExactContact: Boolean) {
         wasEditLaunched = true
         duplicateInitialized = false
-        editContact(contact)
+        if (editExactContact) {
+            editContact(contact)
+        } else {
+            editContact(contact, config.mergeDuplicateContacts)
+        }
     }
 
     private fun openWith() {
@@ -626,7 +628,7 @@ class ViewContactActivity : ContactActivity() {
                     binding.contactSourcesHolder.addView(root)
 
                     contactSource.setOnClickListener {
-                        launchEditContact(key)
+                        launchEditContact(key, true)
                     }
 
                     if (value.lowercase(Locale.getDefault()) == WHATSAPP) {
@@ -819,21 +821,11 @@ class ViewContactActivity : ContactActivity() {
         }
     }
 
-    private fun getDuplicateContacts(callback: () -> Unit) {
-        ContactsHelper(this).getDuplicatesOfContact(contact!!, false) { contacts ->
-            ensureBackgroundThread {
-                duplicateContacts.clear()
-                val displayContactSources = getVisibleContactSources()
-                contacts.filter { displayContactSources.contains(it.source) }.forEach {
-                    val duplicate = ContactsHelper(this).getContactWithId(it.id, it.isPrivate())
-                    if (duplicate != null) {
-                        duplicateContacts.add(duplicate)
-                    }
-                }
-
-                runOnUiThread {
-                    callback()
-                }
+    private fun initializeDuplicateContacts(callback: () -> Unit) {
+        getDuplicateContacts(contact!!, false) {
+            duplicateContacts = it
+            runOnUiThread {
+                callback()
             }
         }
     }